SubContainer finished
- added ability to make subContainers
- added asmdef
- removed unused error references
