+++ Assets/OCNContainer/InternalData/Abstractions/IScopeRegistration.cs
@@ -2,811 +2,16 @@
 
 namespace OCNContainer
 {
-    public interface IScopeRegistration : IScope
+    public partial interface IScopeRegistration
     {
         public void RegisterSubContainer<T>(Action<IScopeRegistration> subContainer);
 
-        public void Register<T1>()
-            where T1 : class, new();
-
-        public void Register<T1, T2>()
-            where T1 : class, new()
-            where T2 : class, new();
-
-        public void Register<T1, T2, T3>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new();
-
-        public void Register<T1, T2, T3, T4>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new();
+        public void Register<T1>() where T1 : class, new();
         
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new();
+        public void RegisterFromInstance<T1>( T1 instance1) where T1 : class;
         
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new();
+        public void Bind<TImplementation, TInterface>() where TImplementation : class, TInterface, new();
         
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24, T25>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new()
-            where T25 : class, new();
-
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24, T25, T26,
-            T27>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new()
-            where T25 : class, new()
-            where T26 : class, new()
-            where T27 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24, T25, T26,
-            T27, T28>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new()
-            where T25 : class, new()
-            where T26 : class, new()
-            where T27 : class, new()
-            where T28 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24, T25, T26,
-            T27, T28, T29>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new()
-            where T25 : class, new()
-            where T26 : class, new()
-            where T27 : class, new()
-            where T28 : class, new()
-            where T29 : class, new();
-        
-        public void Register<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16, T17, T18, T19, T20, T21, T22, T23, T24, T25, T26,
-            T27, T28, T29, T30>()
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new()
-            where T16 : class, new()
-            where T17 : class, new()
-            where T18 : class, new()
-            where T19 : class, new()
-            where T20 : class, new()
-            where T21 : class, new()
-            where T22 : class, new()
-            where T23 : class, new()
-            where T24 : class, new()
-            where T25 : class, new()
-            where T26 : class, new()
-            where T27 : class, new()
-            where T28 : class, new()
-            where T29 : class, new()
-            where T30 : class, new();
-
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10,
-            T11 instance11,
-            T12 instance12,
-            T13 instance13,
-            T14 instance14,
-            T15 instance15)
-            where T1 : class, new()
-            where T2 : class, new()
-            where T3 : class, new()
-            where T4 : class, new()
-            where T5 : class, new()
-            where T6 : class, new()
-            where T7 : class, new()
-            where T8 : class, new()
-            where T9 : class, new()
-            where T10 : class, new()
-            where T11 : class, new()
-            where T12 : class, new()
-            where T13 : class, new()
-            where T14 : class, new()
-            where T15 : class, new();
-
-        public void RegisterFromInstance<T1>(
-            T1 instance1)
-            where T1 : class;
-        
-        public void RegisterFromInstance<T1, T2>(
-            T1 instance1,
-            T2 instance2)
-            where T1 : class
-            where T2 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3)
-            where T1 : class
-            where T2 : class
-            where T3 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class;
-
-        public void RegisterFromInstance<T1, T2, T3, T4, T5>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class
-            where T10 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10,
-            T11 instance11)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class
-            where T10 : class
-            where T11 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10,
-            T11 instance11,
-            T12 instance12)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class
-            where T10 : class
-            where T11 : class
-            where T12 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10,
-            T11 instance11,
-            T12 instance12,
-            T13 instance13)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class
-            where T10 : class
-            where T11 : class
-            where T12 : class
-            where T13 : class;
-        
-        public void RegisterFromInstance<T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14>(
-            T1 instance1,
-            T2 instance2,
-            T3 instance3,
-            T4 instance4,
-            T5 instance5,
-            T6 instance6,
-            T7 instance7,
-            T8 instance8,
-            T9 instance9,
-            T10 instance10,
-            T11 instance11,
-            T12 instance12,
-            T13 instance13,
-            T14 instance14)
-            where T1 : class
-            where T2 : class
-            where T3 : class
-            where T4 : class
-            where T5 : class
-            where T6 : class
-            where T7 : class
-            where T8 : class
-            where T9 : class
-            where T10 : class
-            where T11 : class
-            where T12 : class
-            where T13 : class
-            where T14 : class;
-
-        
-        public void Bind<TImplementation, TInterface>(bool isLazy = false)
-            where TImplementation : class, TInterface, new();
-        
-        public void Bind<TImplementation, TInterface1, TInterface2>(bool isLazy = false)
-            where TImplementation : class, TInterface1, TInterface2, new();
-
-        public void Bind<TImplementation, TInterface1, TInterface2, TInterface3>(bool isLazy = false)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, new();
-        
-        public void Bind<TImplementation, TInterface1, TInterface2, TInterface3, TInterface4>(bool isLazy = false)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, TInterface4, new();
-
-        public void Bind<TImplementation, TInterface1, TInterface2, TInterface3, TInterface4, TInterface5>(bool isLazy = false)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, TInterface4, TInterface5, new();
-
-        public void BindFromInstance<TImplementation, TInterface>(TImplementation instance)
-            where TImplementation : class, TInterface, new();
-        
-        public void BindFromInstance<TImplementation, TInterface1, TInterface2>(TImplementation instance)
-            where TImplementation : class, TInterface1, TInterface2, new();
-        
-        public void BindFromInstance<TImplementation, TInterface1, TInterface2, TInterface3>(TImplementation instance)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, new();
-        
-        public void BindFromInstance<TImplementation, TInterface1, TInterface2, TInterface3, TInterface4>(TImplementation instance)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, TInterface4, new();
-        
-        public void BindFromInstance<TImplementation, TInterface1, TInterface2, TInterface3, TInterface4, TInterface5>(TImplementation instance)
-            where TImplementation : class, TInterface1, TInterface2, TInterface3, TInterface4, TInterface5, new();
-
+        public void BindFromInstance<TImplementation, TInterface>(TImplementation instance) where TImplementation : class, TInterface, new();
     }
 }
